se cambio de lugar donde se llamaba la funcion de paginar, se cambio al archivo App.tsx, se arreglaron los botones con sus respectivos estilos de CSS
